refactor(modularity): finalize crate imports and remove legacy shim modules
