+++ crates/microclaw-app/Cargo.toml
@@ -7,6 +7,7 @@ license = "MIT"
 [dependencies]
 anyhow = "1"
 chrono = { version = "0.4", features = ["serde"] }
+include_dir = "0.7"
 reqwest = { version = "0.12", features = ["json", "blocking", "multipart"] }
 serde_json = "1"
 tracing = "0.1"
