+++ crates/microclaw-app/src/lib.rs
@@ -1,4 +1,5 @@
 //! Application-level runtime and CLI orchestration for MicroClaw.
 
+pub mod builtin_skills;
 pub mod logging;
 pub mod transcribe;
