+++ crates/microclaw-channels/src/lib.rs
@@ -2,3 +2,4 @@
 
 pub mod channel;
 pub mod channel_adapter;
+pub mod delivery;
