+++ crates/microclaw-tools/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod command_runner;
 pub mod path_guard;
+pub mod runtime;
 pub mod sandbox;
 pub mod todo_store;
 pub mod types;
