--- src/db.rs
@@ -1,1 +1,0 @@
-pub use microclaw_storage::db::*;
