--- src/error.rs
@@ -1,1 +1,0 @@
-pub use microclaw_core::error::*;
