+++ src/lib.rs
@@ -1,30 +1,31 @@
 pub mod agent_engine;
-pub mod builtin_skills;
-pub mod channel;
-pub mod channel_adapter;
 pub mod channels;
 pub mod codex_auth;
 pub mod config;
-pub mod db;
 pub mod doctor;
 pub mod embedding;
-pub mod error;
 pub mod gateway;
 pub mod llm;
-pub mod llm_types;
-pub mod logging;
 pub mod mcp;
-pub mod memory;
-pub mod memory_quality;
 pub mod runtime;
-pub mod sandbox;
 pub mod scheduler;
 pub mod setup;
 pub mod skills;
-pub(crate) mod text;
 pub mod tools;
-pub mod transcribe;
 pub mod usage;
 pub mod web;
+
 pub use channels::discord;
 pub use channels::telegram;
+pub use microclaw_app::builtin_skills;
+pub use microclaw_app::logging;
+pub use microclaw_app::transcribe;
+pub use microclaw_channels::channel;
+pub use microclaw_channels::channel_adapter;
+pub use microclaw_core::error;
+pub use microclaw_core::llm_types;
+pub use microclaw_core::text;
+pub use microclaw_storage::db;
+pub use microclaw_storage::memory;
+pub use microclaw_storage::memory_quality;
+pub use microclaw_tools::sandbox;
