--- src/memory.rs
@@ -1,1 +1,0 @@
-pub use microclaw_storage::memory::*;
