--- src/sandbox.rs
@@ -1,1 +1,0 @@
-pub use microclaw_tools::sandbox::*;
