+++ src/scheduler.rs
@@ -6,14 +6,14 @@ use tracing::{error, info};
 
 use crate::agent_engine::process_with_agent;
 use crate::agent_engine::AgentRequestContext;
-use crate::channel::{
+use crate::runtime::AppState;
+use crate::{db::Memory, memory_quality};
+use microclaw_channels::channel::{
     deliver_and_store_bot_message, get_chat_routing, ChatRouting, ConversationKind,
 };
-use crate::db::call_blocking;
-use crate::llm_types::{Message, MessageContent, ResponseContentBlock};
-use crate::runtime::AppState;
-use crate::text::floor_char_boundary;
-use crate::{db::Memory, memory_quality};
+use microclaw_core::llm_types::{Message, MessageContent, ResponseContentBlock};
+use microclaw_core::text::floor_char_boundary;
+use microclaw_storage::db::call_blocking;
 
 pub fn spawn_scheduler(state: Arc<AppState>) {
     tokio::spawn(async move {
