+++ src/setup.rs
@@ -21,8 +21,8 @@ use crate::codex_auth::{
     codex_config_default_openai_base_url, is_openai_codex_provider, provider_allows_empty_api_key,
     resolve_openai_codex_auth,
 };
-use crate::error::MicroClawError;
-use crate::text::floor_char_boundary;
+use microclaw_core::error::MicroClawError;
+use microclaw_core::text::floor_char_boundary;
 
 // ---------------------------------------------------------------------------
 // Declarative channel metadata: adding a new channel only requires adding
