--- src/text.rs
@@ -1,1 +1,0 @@
-pub use microclaw_core::text::*;
