+++ src/tools/memory.rs
@@ -4,9 +4,9 @@ use std::path::PathBuf;
 use std::sync::Arc;
 use tracing::info;
 
-use crate::db::{call_blocking, Database};
-use crate::llm_types::ToolDefinition;
-use crate::memory_quality;
+use microclaw_core::llm_types::ToolDefinition;
+use microclaw_storage::db::{call_blocking, Database};
+use microclaw_storage::memory_quality;
 
 use super::{auth_context_from_input, authorize_chat_access, schema_object, Tool, ToolResult};
 
@@ -213,7 +213,7 @@ mod tests {
     use serde_json::json;
     use std::sync::Arc;
 
-    use crate::db::Database;
+    use microclaw_storage::db::Database;
 
     fn test_dir() -> std::path::PathBuf {
         std::env::temp_dir().join(format!("microclaw_memtool_{}", uuid::Uuid::new_v4()))
