--- src/tools/path_guard.rs
@@ -1,1 +1,0 @@
-pub use microclaw_tools::path_guard::*;
