--- src/tools/web_html.rs
@@ -1,1 +1,0 @@
-pub use microclaw_tools::web_html::*;
