--- src/transcribe.rs
@@ -1,1 +1,0 @@
-pub use microclaw_app::transcribe::*;
